[typos] rename files and change directories for dialogs
