+++ asistente_ladm_col/gui/db_panel/db_schema_db_panel.py
@@ -19,13 +19,11 @@
 from qgis.PyQt.QtWidgets import (QComboBox, QPushButton)
 from qgis.core import (Qgis)
 from qgis.PyQt.QtCore import (Qt,
-                              pyqtSignal,
-                              QCoreApplication,
-                              QTimer)
+                              QCoreApplication)
 from qgis.PyQt.QtGui import QIcon
 
 from .db_config_panel import DbConfigPanel
-from ..dlg_get_db_or_schema_name import DialogGetDBOrSchemaName
+from ...gui.dialogs.dlg_get_db_or_schema_name import DialogGetDBOrSchemaName
 from ...lib.db.db_connector import EnumTestLevel
