+++ asistente_ladm_col/gui/dialogs/dlg_log_quality.py
@@ -21,10 +21,11 @@ from qgis.PyQt.QtWidgets import QDialog
 from qgis.PyQt.QtWidgets import QDialogButtonBox
 from qgis.PyQt.QtCore import (QCoreApplication,
                               QSettings)
-from ..utils import get_ui_class
-from ..utils.qt_utils import save_pdf_format
+from ...utils import get_ui_class
+from ...utils.qt_utils import save_pdf_format
 
-LOG_DIALOG_UI = get_ui_class('dlg_log_quality.ui')
+LOG_DIALOG_UI = get_ui_class('dialogs/dlg_log_quality.ui')
+
 
 class LogQualityDialog(QDialog, LOG_DIALOG_UI):
     def __init__(self, qgis_utils, quality, parent=None):
