+++ asistente_ladm_col/gui/dialogs/dlg_quality.py
@@ -26,19 +26,19 @@ from qgis.PyQt.QtGui import (QBrush,
 from qgis.PyQt.QtWidgets import (QDialog,
                                  QTreeWidgetItem,
                                  QTreeWidgetItemIterator)
-from ..config.general_config import translated_strings
-from ..config.table_mapping_config import (BOUNDARY_POINT_TABLE,
-                                           CONTROL_POINT_TABLE,
-                                           PLOT_TABLE,
-                                           BUILDING_TABLE,
-                                           RIGHT_OF_WAY_TABLE)
-from ..utils import get_ui_class
-from ..resources_rc import *
+from ...config.general_config import translated_strings
+from ...config.table_mapping_config import (BOUNDARY_POINT_TABLE,
+                                            CONTROL_POINT_TABLE,
+                                            PLOT_TABLE,
+                                            BUILDING_TABLE,
+                                            RIGHT_OF_WAY_TABLE)
+from ...utils import get_ui_class
+from ...resources_rc import *
 
-DIALOG_UI = get_ui_class('dlg_quality.ui')
+DIALOG_UI = get_ui_class('dialogs/dlg_quality.ui')
+
 
 class DialogQuality(QDialog, DIALOG_UI):
-
     def __init__(self, db, qgis_utils, quality, parent=None):
         QDialog.__init__(self, parent)
         self.setupUi(self)
