+++ asistente_ladm_col/gui/qgis_model_baker/dlg_import_data.py
@@ -48,7 +48,7 @@ from ...config.general_config import (DEFAULT_EPSG,
                                       CREATE_BASKET_COL,
                                       CREATE_IMPORT_TID,
                                       STROKE_ARCS)
-from ...gui.dlg_get_java_path import DialogGetJavaPath
+from ...gui.dialogs.dlg_get_java_path import DialogGetJavaPath
 from ...utils.qgis_model_baker_utils import get_java_path_from_qgis_model_baker
 from ...utils import get_ui_class
 from ...utils.qt_utils import (Validators,
@@ -61,9 +61,8 @@ from ...lib.db.enum_db_action_type import EnumDbActionType
 
 DIALOG_UI = get_ui_class('qgis_model_baker/dlg_import_data.ui')
 
+
 class DialogImportData(QDialog, DIALOG_UI):
-
-
 
     def __init__(self, iface, db, qgis_utils):
         QDialog.__init__(self)
