+++ asistente_ladm_col/ui/dialogs/dlg_controlled_measurement.ui
