+++ asistente_ladm_col/ui/dialogs/dlg_log_quality.ui
