+++ asistente_ladm_col/ui/dialogs/dlg_settings.ui
